requirements.txt

fastapi
uvicorn[standard]
python-multipart
bcrypt
python-jose[cryptography]
pydantic
Pillow
tflite-runtime

